The NfsPermanentStorage class was moved to the nfs package and your importation in others classes that use it was solved, this was done to organize the permanent storage solutions in unique packages
